backend/requirements.txt
asgiref==3.8.1
Django==5.2
djangorestframework==3.16.0
dotenv==0.9.9
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
